- more comprehensive and better named API for streamlines - removed compilation warnings from header files

+++ include/vtkfigDataSet.h
@@ -364,7 +364,7 @@ namespace vtkfig
     
     int np=pdata->GetNumberOfPoints();
     std::vector<int>pmask(np);
-    for (int i=0;i<pmask.size(); i++) 
+    for (size_t i=0;i<pmask.size(); i++) 
     {
       pmask[i]=-1;
     }
